Refactor directories to more logically group files
